extract `constrain_anon_types` to the `InferCtxt`
No funtional change.
